Adds video_processor that is driven by VRAM and IO
This commit adds a new video_processor struct, that replaces the manual
writing to and binding of textures in App. It updates the textures that
drive the "ui" (should be renamed) shader when vram indicates that the
texture is stale. Certain data like BG scroll and palette is set every
frame as a shader uniform.

In order to achieve this, a memory controller has also been added. This
needs to be fleshed out as the only currently mapped part of the address
bus is 0x8000-0x9FFF and 0xFF40-0xFF4B