Add ALU and registers
This commit adds a 16 bit register, consisting of two 8 bit registers
which can be accessed seperately. These have many helper functions as
part of the Register trait which are used by the ALU.

The ALU is going to be the main place for arithmetic operations to be
modelled in the CPU. The CPU should never modify registers directly
aside from setting values from memory.